requirements.txt

# Core API
fastapi==0.109.1
uvicorn[standard]==0.23.2

# Frontend UI
streamlit==1.28.0
requests==2.32.0

# LLM / RAG
langchain==1.0.0
openai==1.52.0

# Vector Database
faiss-cpu==1.8.0

# Utilities
numpy==1.26.0
pydantic==2.8.0
python-dotenv==1.1.0
python-multipart==0.0.6

# Testing
pytest==8.3.3

# Optional for monitoring/logging
loguru==0.7.0
prometheus-client==0.22.0
